+++ TransitionMatrix/TransitionMatrix.h
@@ -54,9 +54,9 @@ public:
     int cols() const;
     long double nonZeros() const;
 
-    void write(ofstream &fd, SparseMatrix<int, RowMajor> *tmp_result);
-    void writeCondensed(ofstream &fd, Eigen::SparseMatrix<int, RowMajor> *tmp_result);
-    void writeForPagerank(ofstream &fd, Eigen::SparseMatrix<int, RowMajor> *tmp_result);
+    void write(ofstream &fd, SparseMatrix<double, RowMajor> *tmp_result);
+    void writeCondensed(ofstream &fd, Eigen::SparseMatrix<double, RowMajor> *tmp_result);
+    void writeForPagerank(ofstream &fd, Eigen::SparseMatrix<double, RowMajor> *tmp_result);
 };
